reorganize tests according to the new contract structure
