+++ test/proxy/beacon/BeaconProxy.test.js
