+++ test/token/ERC1155/presets/ERC1155PresetMinterPauser.test.js
