+++ test/token/ERC20/extensions/ERC20Burnable.behavior.js
