+++ test/token/ERC20/extensions/ERC20Burnable.test.js
@@ -1,6 +1,6 @@
 const { BN } = require('@openzeppelin/test-helpers');
 
-const { shouldBehaveLikeERC20Burnable } = require('./behaviors/ERC20Burnable.behavior');
+const { shouldBehaveLikeERC20Burnable } = require('./ERC20Burnable.behavior');
 const ERC20BurnableMock = artifacts.require('ERC20BurnableMock');
 
 contract('ERC20Burnable', function (accounts) {
