+++ test/token/ERC20/extensions/ERC20Capped.behavior.js
