+++ test/token/ERC20/extensions/ERC20Snapshot.test.js
