+++ test/token/ERC20/presets/ERC20PresetFixedSupply.test.js
