+++ test/token/ERC721/presets/ERC721PresetMinterPauserAutoId.test.js
