+++ test/token/ERC777/presets/ERC777PresetFixedSupply.test.js
