+++ test/utils/cryptography/MerkleProof.test.js
@@ -1,6 +1,6 @@
 require('@openzeppelin/test-helpers');
 
-const { MerkleTree } = require('../helpers/merkleTree.js');
+const { MerkleTree } = require('../../helpers/merkleTree.js');
 const { keccakFromString, bufferToHex } = require('ethereumjs-util');
 
 const { expect } = require('chai');
