+++ test/utils/cryptography/draft-EIP712.test.js
@@ -1,7 +1,7 @@
 const ethSigUtil = require('eth-sig-util');
 const Wallet = require('ethereumjs-wallet').default;
 
-const { EIP712Domain, domainSeparator } = require('../helpers/eip712');
+const { EIP712Domain, domainSeparator } = require('../../helpers/eip712');
 
 const EIP712 = artifacts.require('EIP712External');
