+++ test/utils/introspection/SupportsInterface.behavior.js
