+++ test/utils/math/SafeCast.test.js
